refactor: align terminology — decode/decoding for bytecode, X2Y for rewrite
