+++ docs/superpowers/specs/2026-04-21-bytecode-to-ssa-lowering.md
@@ -124,7 +124,7 @@ This draws on categorical / substructural IRs (Joyal-Street string diagrams, Sel
 1. **Atom non-cloning as a first-class invariant.** Atoms physically cannot be duplicated. An explicit `Dup` statement in the IR is exactly the right place for a later linear-type pass to check: any `Dup` whose operand has an atom-bearing type is a physical error. The mainstream vanishing-stack-ops treatment would have to reconstruct this information from alias analysis or equivalent.
 2. **Structural round-trippability to bytecode.** Every bytecode instruction has exactly one corresponding `stack_move` statement. Compiling Kirin IR back down to bytecode is a structural inverse of decoding — no reconstruction of missing stack ops.
 
-The cost is that stack ops have to be explicitly collapsed when lowering `stack_move` into the downstream target dialects (`move`, `ilist`, etc.): `Dup` and `Swap` map their result SSA values back to their inputs, `Pop` produces nothing. This is a trivial per-statement transformation in `lower_stack_move` — but it's a step that the mainstream treatment avoids by never introducing the statements in the first place.
+The cost is that stack ops have to be explicitly collapsed when lowering `stack_move` into the downstream target dialects (`move`, `ilist`, etc.): `Dup` and `Swap` map their result SSA values back to their inputs, `Pop` produces nothing. This is a trivial per-statement transformation in `stack_move2move` — but it's a step that the mainstream treatment avoids by never introducing the statements in the first place.
 
 ## Where the technique lives: decoder vs rewrite pass
 
@@ -134,7 +134,7 @@ The Bloqade Lanes implementation puts the stack-to-SSA conversion **in the decod
 
 **Stack ops are preserved as explicit SSA statements** rather than being eliminated invisibly during decoding. `Dup` is an SSA identity statement (`%y = Dup %x`, semantically `%y ≡ %x`); `Swap` is an SSA permutation statement (`%a, %b = Swap %x, %y`, with `%a ≡ %y` and `%b ≡ %x`); `Pop` is an SSA discard statement (`Pop %x`, no result). This is the "linear IR" style described earlier — it preserves the bytecode's structural operations at the SSA level so that downstream analyses (notably a future atom-non-cloning type check) can attach invariants to them, and so that lowering from Kirin *back* to bytecode is a structural inverse of decoding.
 
-The follow-up rewrite `lower_stack_move` becomes a **mechanical per-statement translation** from `stack_move` into target dialects (`move`, `ilist`, `py.constant`, `py.indexing`, `annotate`, `func`). `Dup` and `Swap` collapse during that rewrite: their result SSA values are mapped to the same target-dialect SSA values as their inputs. `Pop` produces no target statement. The linear-IR shape lives only in `stack_move`; the target IR is conventional SSA.
+The follow-up rewrite `stack_move2move` becomes a **mechanical per-statement translation** from `stack_move` into target dialects (`move`, `ilist`, `py.constant`, `py.indexing`, `annotate`, `func`). `Dup` and `Swap` collapse during that rewrite: their result SSA values are mapped to the same target-dialect SSA values as their inputs. `Pop` produces no target statement. The linear-IR shape lives only in `stack_move`; the target IR is conventional SSA.
 
 Conceptual shape of the decoder:
 
@@ -189,7 +189,7 @@ The virtual stack is a `list[ir.SSAValue]` on the decoder instance. `DecodeError
 
 **Compatibility with future control flow.** This decoder design composes cleanly with Approach 3 (block-argument SSA, see §"Unifying with the host dataflow framework" below) when branching is added: the virtual stack at a block exit becomes the argument list for the successor block, and the explicit `stack_move.Dup`/`Swap`/`Pop` statements operate uniformly on both block-argument SSA values and locally produced SSA values — they don't care which kind of SSA value they're manipulating.
 
-**Toward a bytecode decoder framework (planned follow-up).** The v1 implementation is deliberately monolithic — one `stack_move` dialect, one `BytecodeDecoder`, one `lower_stack_move` rewrite — so the end-to-end mechanics can be validated on a small, reviewable surface. Once those boundaries are stable, the natural next step is to decompose the work into a **bytecode decoder framework**: instruction families become sub-dialects (stack ops, constants, atom ops, gates, measurement, arrays, annotations, control flow), each sub-dialect registers a **decoding handler** that the top-level decoder dispatches to for its opcodes, and each sub-dialect also owns a **rewrite chunk** that consumes its own statements and emits into a target dialect. The virtual stack threads through the decoder's chain of handlers as shared state. The payoff is independently testable, replaceable, reusable pieces — and a framework that can handle new stack-based bytecode variants by swapping sub-dialect handlers rather than rewriting the whole decoder. This is post-v1 work but it's the direction this prototype is paving toward.
+**Toward a bytecode decoder framework (planned follow-up).** The v1 implementation is deliberately monolithic — one `stack_move` dialect, one `BytecodeDecoder`, one `stack_move2move` rewrite — so the end-to-end mechanics can be validated on a small, reviewable surface. Once those boundaries are stable, the natural next step is to decompose the work into a **bytecode decoder framework**: instruction families become sub-dialects (stack ops, constants, atom ops, gates, measurement, arrays, annotations, control flow), each sub-dialect registers a **decoding handler** that the top-level decoder dispatches to for its opcodes, and each sub-dialect also owns a **rewrite chunk** that consumes its own statements and emits into a target dialect. The virtual stack threads through the decoder's chain of handlers as shared state. The payoff is independently testable, replaceable, reusable pieces — and a framework that can handle new stack-based bytecode variants by swapping sub-dialect handlers rather than rewriting the whole decoder. This is post-v1 work but it's the direction this prototype is paving toward.
 
 ## Error conditions
