+++ python/tests/rewrite/test_stack_move2move.py
@@ -8,7 +8,7 @@ from bloqade.lanes.layout.encoding import (
     LocationAddress as EncodingLocationAddress,
     ZoneAddress as EncodingZoneAddress,
 )
-from bloqade.lanes.rewrite.lower_stack_move import LowerStackMove
+from bloqade.lanes.rewrite.stack_move2move import RewriteStackMoveToMove
 
 
 def _build_stack_move_block(stmts: list[ir.Statement]) -> ir.Block:
@@ -32,7 +32,7 @@ def _build_stack_move_block(stmts: list[ir.Statement]) -> ir.Block:
 def test_empty_block_emits_load_and_func_return():
     # _build_stack_move_block synthesises a trailing ConstInt(0) + Return.
     block = _build_stack_move_block([])
-    result = Walk(LowerStackMove()).rewrite(block)
+    result = Walk(RewriteStackMoveToMove()).rewrite(block)
     assert result.has_done_something
     # Expect a move.Load at block start and a func.Return; the stack_move
     # Return should have been deleted.
@@ -43,7 +43,7 @@ def test_empty_block_emits_load_and_func_return():
 def test_const_float_emits_py_constant_and_tracks_value():
     cf = stack_move.ConstFloat(value=1.5)
     block = _build_stack_move_block([cf])
-    rule = LowerStackMove()
+    rule = RewriteStackMoveToMove()
     Walk(rule).rewrite(block)
     # py.Constant statement emitted with value 1.5.
     py_const = next(s for s in block.stmts if isinstance(s, py.Constant))
@@ -58,7 +58,7 @@ def test_const_loc_tracks_attribute_value():
     addr = LocationAddress(0, 0, 0)
     cl = stack_move.ConstLoc(value=addr)
     block = _build_stack_move_block([cl])
-    rule = LowerStackMove()
+    rule = RewriteStackMoveToMove()
     Walk(rule).rewrite(block)
     # The stack_move SSA is mapped to its raw attribute (for lifting into
     # downstream move.* attributes).
@@ -69,7 +69,7 @@ def test_pop_is_dropped():
     cf = stack_move.ConstFloat(value=1.0)
     pop = stack_move.Pop(value=cf.result)
     block = _build_stack_move_block([cf, pop])
-    Walk(LowerStackMove()).rewrite(block)
+    Walk(RewriteStackMoveToMove()).rewrite(block)
     # No target statement for Pop, and the original stack_move.Pop is gone.
     assert not any(isinstance(s, stack_move.Pop) for s in block.stmts)
 
@@ -80,7 +80,7 @@ def test_dup_redirects_uses_to_input():
     # Downstream consumer that references Dup's result.
     consumer = stack_move.Pop(value=dup.result)
     block = _build_stack_move_block([cf, dup, consumer])
-    Walk(LowerStackMove()).rewrite(block)
+    Walk(RewriteStackMoveToMove()).rewrite(block)
     # Dup is gone; Pop is also lowered away.
     assert not any(isinstance(s, stack_move.Dup) for s in block.stmts)
     assert not any(isinstance(s, stack_move.Pop) for s in block.stmts)
@@ -95,7 +95,7 @@ def test_swap_permutes_uses():
     p_top = stack_move.Pop(value=sw.out_top)
     p_bot = stack_move.Pop(value=sw.out_bot)
     block = _build_stack_move_block([a, b, sw, p_top, p_bot])
-    Walk(LowerStackMove()).rewrite(block)
+    Walk(RewriteStackMoveToMove()).rewrite(block)
     # Swap is gone.
     assert not any(isinstance(s, stack_move.Swap) for s in block.stmts)
 
@@ -107,7 +107,7 @@ def test_fill_lowers_to_move_fill_with_attribute_locations():
     cl1 = stack_move.ConstLoc(value=a1)
     fill = stack_move.Fill(locations=(cl0.result, cl1.result))
     block = _build_stack_move_block([cl0, cl1, fill])
-    Walk(LowerStackMove()).rewrite(block)
+    Walk(RewriteStackMoveToMove()).rewrite(block)
     mf = next(s for s in block.stmts if isinstance(s, move.Fill))
     # stack_move.ConstLoc now stores encoding-layer LocationAddress values
     # directly (matching the move dialect convention), so the rewrite just
@@ -128,7 +128,7 @@ def test_local_r_lowers_with_attribute_lifting():
         locations=(cl.result,),
     )
     block = _build_stack_move_block([cf_theta, cf_phi, cl, lr])
-    Walk(LowerStackMove()).rewrite(block)
+    Walk(RewriteStackMoveToMove()).rewrite(block)
     mr = next(s for s in block.stmts if isinstance(s, move.LocalR))
     # move.LocalR stores rotation angles as SSA values; axis_angle and
     # rotation_angle pass through from the stack_move statement. The SSA
@@ -148,7 +148,7 @@ def test_local_rz_lowers_with_attribute_lifting():
     cl = stack_move.ConstLoc(value=LocationAddress(0, 1, 0))
     lr = stack_move.LocalRz(rotation_angle=cf_theta.result, locations=(cl.result,))
     block = _build_stack_move_block([cf_theta, cl, lr])
-    Walk(LowerStackMove()).rewrite(block)
+    Walk(RewriteStackMoveToMove()).rewrite(block)
     mr = next(s for s in block.stmts if isinstance(s, move.LocalRz))
     rot_const = mr.rotation_angle.owner
     assert isinstance(rot_const, py.Constant)
@@ -162,7 +162,7 @@ def test_global_r_lowers_with_attribute_lifting():
     cf_phi = stack_move.ConstFloat(value=0.5)
     gr = stack_move.GlobalR(axis_angle=cf_phi.result, rotation_angle=cf_theta.result)
     block = _build_stack_move_block([cf_theta, cf_phi, gr])
-    Walk(LowerStackMove()).rewrite(block)
+    Walk(RewriteStackMoveToMove()).rewrite(block)
     mr = next(s for s in block.stmts if isinstance(s, move.GlobalR))
     axis_const = mr.axis_angle.owner
     rot_const = mr.rotation_angle.owner
@@ -176,7 +176,7 @@ def test_global_rz_lowers_with_attribute_lifting():
     cf_theta = stack_move.ConstFloat(value=0.6)
     gr = stack_move.GlobalRz(rotation_angle=cf_theta.result)
     block = _build_stack_move_block([cf_theta, gr])
-    Walk(LowerStackMove()).rewrite(block)
+    Walk(RewriteStackMoveToMove()).rewrite(block)
     mr = next(s for s in block.stmts if isinstance(s, move.GlobalRz))
     rot_const = mr.rotation_angle.owner
     assert isinstance(rot_const, py.Constant)
@@ -189,9 +189,9 @@ def test_cz_lowers_with_attribute_zone():
     cz_zone = stack_move.ConstZone(value=ZoneAddress(0))
     cz = stack_move.CZ(zone=cz_zone.result)
     block = _build_stack_move_block([cz_zone, cz])
-    Walk(LowerStackMove()).rewrite(block)
+    Walk(RewriteStackMoveToMove()).rewrite(block)
     mcz = next(s for s in block.stmts if isinstance(s, move.CZ))
-    # LowerStackMove wraps native ZoneAddress into the encoding wrapper.
+    # RewriteStackMoveToMove wraps native ZoneAddress into the encoding wrapper.
     assert mcz.zone_address == EncodingZoneAddress(0)
 
 
@@ -200,7 +200,7 @@ def test_measure_single_zone_emits_single_zone_measure():
     cl1 = stack_move.ConstLoc(value=LocationAddress(0, 1, 0))
     m = stack_move.Measure(locations=(cl0.result, cl1.result))
     block = _build_stack_move_block([cl0, cl1, m])
-    Walk(LowerStackMove()).rewrite(block)
+    Walk(RewriteStackMoveToMove()).rewrite(block)
     mm = next(s for s in block.stmts if isinstance(s, move.Measure))
     # One zone (both locs are in zone 0).
     assert len(mm.zones) == 1
@@ -213,7 +213,7 @@ def test_measure_multi_zone_dedups():
     cl2 = stack_move.ConstLoc(value=LocationAddress(0, 1, 0))
     m = stack_move.Measure(locations=(cl0.result, cl1.result, cl2.result))
     block = _build_stack_move_block([cl0, cl1, cl2, m])
-    Walk(LowerStackMove()).rewrite(block)
+    Walk(RewriteStackMoveToMove()).rewrite(block)
     mm = next(s for s in block.stmts if isinstance(s, move.Measure))
     assert len(mm.zones) == 2
 
@@ -225,7 +225,7 @@ def test_await_measure_lowers_without_error():
     m = stack_move.Measure(locations=(cl.result,))
     aw = stack_move.AwaitMeasure(future=m.result)
     block = _build_stack_move_block([cl, m, aw])
-    Walk(LowerStackMove()).rewrite(block)  # should not raise
+    Walk(RewriteStackMoveToMove()).rewrite(block)  # should not raise
 
 
 def test_new_array_lowers_to_ilist_new():
@@ -233,7 +233,7 @@ def test_new_array_lowers_to_ilist_new():
 
     na = stack_move.NewArray(type_tag=0, dim0=4, dim1=0)
     block = _build_stack_move_block([na])
-    Walk(LowerStackMove()).rewrite(block)
+    Walk(RewriteStackMoveToMove()).rewrite(block)
     assert any(isinstance(s, ilist.New) for s in block.stmts)
 
 
@@ -243,7 +243,7 @@ def test_set_detector_lowers_to_annotate():
     na = stack_move.NewArray(type_tag=0, dim0=1, dim1=0)
     sd = stack_move.SetDetector(array=na.result)
     block = _build_stack_move_block([na, sd])
-    Walk(LowerStackMove()).rewrite(block)
+    Walk(RewriteStackMoveToMove()).rewrite(block)
     assert any(isinstance(s, annotate.stmts.SetDetector) for s in block.stmts)
 
 
@@ -253,5 +253,5 @@ def test_set_observable_lowers_to_annotate():
     na = stack_move.NewArray(type_tag=0, dim0=1, dim1=0)
     so = stack_move.SetObservable(array=na.result)
     block = _build_stack_move_block([na, so])
-    Walk(LowerStackMove()).rewrite(block)
+    Walk(RewriteStackMoveToMove()).rewrite(block)
     assert any(isinstance(s, annotate.stmts.SetObservable) for s in block.stmts)
